+++ internal/codeintel/testdata/sample.go
@@ -1,0 +1,12 @@
+package sample
+
+import "fmt"
+
+type Foo struct{ X int }
+
+func (f *Foo) Bar() int { return f.X }
+
+func Run() {
+	f := &Foo{X: 1}
+	fmt.Println(f.Bar())
+}